refactor: 불필요한 props 전달, 타입 선언 중복 해결
1. 불필요한 props 전달
	- 기존: Main 컴포넌트에서 Project, Activity 컴포넌트 호출 시 불필요한 props를 전달	- 해결: props 전달을 삭제하고, 호출된 컴포넌트에서 직접 상수를 호출(SectionTypeConstants.ts)하는 방식으로 변경
	- 변경 파일 목록:
		Main.tsx, Section.tsx, Project.tsx, Activity.tsx, SectionTypeConstants.ts(신규 생성)

2. 타입 선언 중복
	(a. Main이 전달할 ref 타입)
	- 기존: Project, Activity 컴포넌트에서 각각 Article 컴포넌트에 전달할 ref 타입을 선언
	- 해결: Main 컴포넌트에 전달받은 ref타입을 선언하기로 하고, export 하여 자식 컴포넌트에서 사용
	- 변경 파일 목록:
		Main.tsx, Project.tsx, Activity.tsx

	(b. Project, Activity가 전달할 props 타입)
	- 기존: Project, Activity 컴포넌트에서 각각 Article 컴포넌트에 전달할 props 타입을 선언
	- 해결: Article 컴포넌트에 전달받은 props 타입을 선언하기로 하고, export 하여 부모 컴포넌트에서 사용
	- 변경 파일 목록:
		Project.tsx, Activity.tsx, SectionTitle.tsx
