+++ src/constants/SectionTypeConstants.ts
@@ -1,0 +1,4 @@
+export const TYPE = {
+    PROJECT: 'PROJECT',
+    ACTIVITY: 'ACTIVITY',
+};
